Add a public api to ensureconda
