+++ ensureconda/__init__.py
@@ -1,1 +1,4 @@
 __version__ = "1.0.0"
+
+
+from .api import ensureconda
