+++ ensureconda/cli.py
@@ -1,22 +1,9 @@
-import contextlib
-import os
-import pathlib
-import platform
-import shutil
-import stat
 import sys
 import time
-import typing
-from os import PathLike
-from pathlib import Path
-from typing import IO, Iterator, Optional, Union
 
-import appdirs
 import click
-import requests
 
-is_windows = platform.system() == "Windows"
-site_path = pathlib.Path(appdirs.user_data_dir("ensure-conda"))
+from ensureconda.api import ensureconda
 
 
 @click.command(help="Ensures that a conda/mamba is installed.")
@@ -36,7 +23,7 @@ site_path = pathlib.Path(appdirs.user_data_dir("ensure-conda"))
 def ensureconda_cli(mamba, micromamba, conda, conda_exe, no_install):
     # We run the loop twice, once to find all the eligible condas without installation
     # and once if you haven't found anything after installation
-    exe = ensure_conda(
+    exe = ensureconda(
         mamba=mamba,
         micromamba=micromamba,
         conda=conda,
@@ -44,7 +31,7 @@ def ensureconda_cli(mamba, micromamba, conda, conda_exe, no_install):
         no_install=True,
     )
     if not exe and not no_install:
-        exe = ensure_conda(
+        exe = ensureconda(
             mamba=mamba,
             micromamba=micromamba,
             conda=conda,
@@ -60,173 +47,6 @@ def ensureconda_cli(mamba, micromamba, conda, conda_exe, no_install):
     else:
         print("Could not find compatible executable", file=sys.stderr)
         sys.exit(1)
-
-
-def ext_candidates(fpath: str) -> Iterator[str]:
-    if is_windows:
-        exts = os.environ.get("PATHEXT", "").split(os.pathsep)
-    else:
-        exts = ["", ".exe"]
-    for ext in exts:
-        yield fpath + ext
-
-
-def is_exe(fpath: Optional[Path]):
-    if fpath is None:
-        return False
-    return os.path.exists(fpath) and os.access(fpath, os.X_OK) and os.path.isfile(fpath)
-
-
-def which_no_shims(executable: str) -> Optional[str]:
-    """Ensure that environment manager tools like pyenv don't break us
-
-    Certain tools like pyenv make use of magic shim directories that contain
-    executables that are unsuitable for our use case.
-
-    These executables will exist but may fail on use due to not being in the right
-    pyenv execution environment.
-    """
-    path_list = [
-        p
-        for p in os.environ["PATH"].split(os.pathsep)
-        if os.path.join(".pyenv", "shims") not in p
-    ]
-    return shutil.which(executable, path=os.pathsep.join(path_list))
-
-
-def resolve_executable(
-    exe_name: str, path_prefix: Optional[Path] = site_path
-) -> Iterator[Path]:
-    for candidate in ext_candidates(exe_name):
-        if path_prefix is not None:
-            prefixed_exe = path_prefix / candidate
-            if is_exe(prefixed_exe):
-                yield prefixed_exe
-
-        # which based exe
-        exe = which_no_shims(candidate)
-        if exe:
-            exe_path = pathlib.Path(exe)
-            if is_exe(exe_path):
-                yield exe_path
-
-
-def conda_executables() -> Iterator[PathLike]:
-    conda_exe_from_env = os.environ.get("CONDA_EXE")
-    if conda_exe_from_env:
-        if is_exe(Path(conda_exe_from_env)):
-            yield Path(conda_exe_from_env)
-    yield from resolve_executable("conda")
-
-
-def conda_standalone_executables() -> Iterator[PathLike]:
-    yield from resolve_executable("conda_standalone")
-
-
-def mamba_executables() -> Iterator[PathLike]:
-    yield from resolve_executable("mamba")
-
-
-def micromamba_executables() -> Iterator[PathLike]:
-    yield from resolve_executable("micromamba")
-
-
-def safe_next(it):
-    try:
-        return next(it)
-    except StopIteration:
-        return None
-
-
-def ensure_conda(
-    *,
-    mamba: bool = True,
-    micromamba: bool = True,
-    conda: bool = True,
-    conda_exe: bool = True,
-    no_install: bool = False,
-) -> Optional[PathLike]:
-    if mamba:
-        exe = safe_next(mamba_executables())
-        if exe:
-            return exe
-    if micromamba:
-        exe = safe_next(micromamba_executables())
-        if exe:
-            return exe
-        if not no_install:
-            return install_micromamba()
-    if conda:
-        exe = safe_next(conda_executables())
-        if exe:
-            return exe
-    if conda_exe:
-        exe = safe_next(conda_standalone_executables())
-        if exe:
-            return exe
-        if not no_install:
-            return install_conda_exe()
-    return None
-
-
-@contextlib.contextmanager
-def new_executable(target_filename: "PathLike") -> Iterator[IO[bytes]]:
-    with open(target_filename, "wb") as fo:
-        yield fo
-    st = os.stat(target_filename)
-    os.chmod(target_filename, st.st_mode | stat.S_IXUSR)
-
-
-def install_conda_exe() -> Optional[Path]:
-    conda_exe_prefix = "https://repo.anaconda.com/pkgs/misc/conda-execs"
-    if platform.system() == "Linux":
-        conda_exe_file = "conda-latest-linux-64.exe"
-    elif platform.system() == "Darwin":
-        conda_exe_file = "conda-latest-osx-64.exe"
-    elif platform.system() == "NT":
-        conda_exe_file = "conda-latest-win-64.exe"
-    else:
-        # TODO: Support other platforms here
-        raise ValueError(f"Unsupported platform: {platform.system()}")
-
-    resp = requests.get(f"{conda_exe_prefix}/{conda_exe_file}", allow_redirects=True)
-    resp.raise_for_status()
-
-    site_path.mkdir(parents=True, exist_ok=True)
-    ext = ".exe" if is_windows else ""
-    target_filename = site_path / f"conda_standalone{ext}"
-    with new_executable(target_filename) as fo:
-        fo.write(resp.content)
-    return pathlib.Path(target_filename)
-
-
-def install_micromamba() -> Optional[Path]:
-    """Install micromamba into the installation"""
-    if platform.system() == "Linux":
-        p = "linux-64"
-    elif platform.system() == "Darwin":
-        p = "osx-64"
-    else:
-        return None
-        raise ValueError(f"Unsupported platform: {platform.system()}")
-    url = f"https://micromamba.snakepit.net/api/micromamba/{p}/latest"
-    resp = requests.get(url, allow_redirects=True)
-    resp.raise_for_status()
-    import tarfile
-    import io
-
-    tarball = io.BytesIO(resp.content)
-    tarball.seek(0)
-    with tarfile.open(mode="r:bz2", fileobj=tarball) as tf:
-        fo = tf.extractfile("bin/micromamba")
-        if fo is None:
-            raise RuntimeError("Could not extract micromamba executable!")
-        ext = ".exe" if is_windows else ""
-        site_path.mkdir(parents=True, exist_ok=True)
-        target_path = site_path / f"micromamba{ext}"
-        with new_executable(target_path) as exe_fo:
-            exe_fo.write(fo.read())
-        return target_path
 
 
 if __name__ == "__main__":
