+++ setup.py
@@ -10,7 +10,7 @@ from setuptools import find_packages, setup
 with open("README.md", "r", encoding="utf-8") as f:
     readme = f.read()
 
-REQUIRES = ["click", "requests", "appdirs"]
+REQUIRES = ["click", "requests", "appdirs", "filelock"]
 
 setup(
     name="ensureconda",
